fix(init): change to holistic ambiguity detection
In the previous implementation, a vision solution candidate was first
checked for the IMU match ambiguity, and was possibly rejected from the
final ambiguity detection. With this, our initializer might incorrectly
accept an ambiguous initialization for instance where there are two
vision solutions A and B, with A ambiguous by two possible scale
solutions and B no ambiguity, and the system is overall ambiguous since
overall three solutions are possible.

This commit corrects this behavior by changing the ambiguity detection
logic to the holistic manner. We first collect all possible vision-IMU
solutions before detecting ambiguity, and reject the solution if the
final number of possible solution pairs is greater than 1.